Refactor financial calculations to use integer-based precision and enhance expense tracking UI.
This commit migrates financial models from `Double` to `Long` to avoid floating-point precision issues and introduces a more detailed breakdown of daily vs. general progress for time-based expenses.

- Refactor `Order`, `TimeBasedExpense`, and related use cases to use `Long` for all currency amounts.
- Rename `GetOrdersNetTotalUseCase` to `GetOrdersAmountAfterKmUseCase` and update its implementation to return totals after mileage deductions.
- Update `ApplyTimeBasedDeductionUseCase` to use `BigDecimal` for fair-share calculations with ceiling rounding to ensure daily goals are met.
- Enhance `TimeBasedExpense` model with methods to calculate both full and remaining daily quotas using rounding logic.
- Expand `TimeBasedExpenseItem` UI and state to display separate progress indicators and goals for "Daily" vs. "General" savings.
- Update `GetTotalTimeBasedExpensesImpactUseCase` to cap impact calculations at the current date.
- Refactor `HomeScreen` and `TotalOrdersAmount` to use stateless content components for better preview support and separation of concerns.
- Update currency formatting across ViewModels to display amounts without decimal fractions (EUR/GERMANY locale).
- Expand `TimeBasedExpesesFake` with realistic yearly expenses (SOAT, RTM).
- Update unit tests to reflect `Long` type changes and verify new deduction result structures.

+++ app/src/main/java/com/control_delivery/finanzas_delivery/data/repository/OrdersFake.kt
@@ -7,42 +7,42 @@ import com.control_delivery.finanzas_delivery.utils.DateUtils
 data class Orders(var orders: MutableList<Order> = mutableListOf(
     Order(platform = "DIDI",
         customerAddress = "123 Main St",
-        totalAmount = 5300.0,
+        totalAmount = 5300,
         status = OrderStatus.DELIVERED,
         timestamp = DateUtils.getTimestampRange("16-02-2026").first),
     Order(platform = "ARMIRENE",
         customerAddress = "456 Elm St",
-        totalAmount = 4700.0,
+        totalAmount = 4700,
         status = OrderStatus.DELIVERED,
         timestamp = DateUtils.getTimestampRange("16-02-2026").first),
     Order(platform = "DIDI",
         customerAddress = "789 Oak St",
-        totalAmount = 3600.0,
+        totalAmount = 3600,
         status = OrderStatus.DELIVERED,
         timestamp = DateUtils.getTimestampRange("16-02-2026").first),
     Order(platform = "RAPPI",
         customerAddress = "321 Pine St",
-        totalAmount = 6850.0,
+        totalAmount = 6850,
         status = OrderStatus.DELIVERED,
         timestamp = DateUtils.getTimestampRange("16-02-2026").first),
     Order(platform = "CABIFY",
         customerAddress = "654 Maple St",
-        totalAmount = 3490.0,
+        totalAmount = 3490,
         status = OrderStatus.ON_THE_WAY_TO_DELIVERY,
         timestamp = DateUtils.getTimestampRange("16-02-2026").first),
     Order(platform = "DIDI",
         customerAddress = "987 Cedar St",
-        totalAmount = 5200.0,
+        totalAmount = 5200,
         status = OrderStatus.DELIVERED,
         timestamp = DateUtils.getTimestampRange("23-02-2026").first),
     Order(platform = "ARMIRENE",
         customerAddress = "159 Birch",
-        totalAmount = 4700.0,
+        totalAmount = 4700,
         status = OrderStatus.DELIVERED,
         timestamp = DateUtils.getTimestampRange("24-02-2026").first),
     Order(platform = "RAPPI",
         customerAddress = "452 Street",
-        totalAmount = 4700.0,
+        totalAmount = 4700,
         status = OrderStatus.DELIVERED,
         timestamp = DateUtils.getTimestampRange("24-02-2026").first),
     )) {
@@ -56,16 +56,17 @@ data class Orders(var orders: MutableList<Order> = mutableListOf(
         }
     }
 
-    fun getOrdersTotalAmount(): Double {
+    fun getOrdersTotalAmount(): Long {
         return orders.filter {
             it.status == OrderStatus.DELIVERED }.sumOf { it.totalAmount }
     }
 
-    fun getOrdersTotalAmount(startDate: Long, endDate: Long): Double {
+    fun getOrdersTotalAmount(startDate: Long, endDate: Long): Long {
         return orders.filter {
             it.status == OrderStatus.DELIVERED && it.timestamp in startDate..endDate}
             .sumOf { it.totalAmount }
         }
+
 
     fun getOrdersByStatesInDateRange(status: List<OrderStatus>,
                                      startDate: Long, endDate: Long): List<Order> {
