+++ app/src/main/java/com/control_delivery/finanzas_delivery/data/repository/TimeBasedExpenseInMemoryRepository.kt
@@ -17,12 +17,13 @@ class TimeBasedExpenseInMemoryRepository(
         return _expensesFlow.asStateFlow()
     }
 
-    override fun getDailyAmountExpenses(today: LocalDate): Flow<Double> {
+    override fun getDailyAmountExpenses(today: LocalDate): Flow<Long> {
         return _expensesFlow.map { list ->
             list.filter { !it.isDeleted }
                 .sumOf { it.getDailyAmount(today) }
         }
     }
+
 
     override fun getExpenseById(id: String): Flow<TimeBasedExpense?> {
         return _expensesFlow.map { expenses ->
